DB

Documentation

deleteOldDb :: IO ()

Check to see if a database file already exists and deletes it

dbConnect :: IO Connection

Create database connection

initialiseDB :: IO ()

Create the events table to hold earthquake records

insertDB :: [Earthquake] -> IO ()

Insert all earthquakes into the database, the close the database

buildRecord :: IConnection conn => conn -> Earthquake -> IO Integer

Convert an Earthquake into an Sql record

getFromDB :: String -> IO [[SqlValue]]

Query the database and return a list of Sql records

getDbContentsAsList :: [[SqlValue]] -> [[String]]

Convert a list of Sql records into a list of Haskell strings

createJson :: [[Char]] -> [[Char]] -> [[Char]] -> [[Char]] -> [[Char]] -> [[Char]] -> [[Char]] -> [[Char]] -> [[Char]] -> [Char]

Creates Json and concatenates magnitude, url, etc to the json object. Create Json is called recursively

getMag :: [a] -> a

Get magnitude from the json

getUrl :: [a] -> a

Get URL from the json

getLat :: [a] -> a

Get latitude from the json

getLong :: [a] -> a

Get longitude from the json

getDepth :: [a] -> a

Get depth from the json

getPlace :: [a] -> a

Get place from the json

getYear :: [a] -> a

Get year from the json

getMonth :: [a] -> a

Get month from the json

getDay :: [a] -> a

Get day from the json

getFullJson :: [[[Char]]] -> [Char]

Create and retrieve full json