Parser

Documentation

getEarthquakes :: Either String String -> [String]

Takes the result of the http download and produces a list of raw earthquake data strings

getProperty :: String -> String -> String

Takes a raw earthquake string and returns a property value

getCoordinates :: String -> [Maybe Double]

Takes a raw earthquake string and returns a list of floats showing coordinates

getUTC :: EpochTime -> CalendarTime

Converts data from Epoch time to UTC time

getYear :: Maybe EpochTime -> Maybe Int

Return the year from Epoch time

getMonth :: Num a => Maybe EpochTime -> Maybe a

Return the month from Epoch time

getDay :: Maybe EpochTime -> Maybe Int

Return the day from Epoch time