Types

Documentation

data Earthquake

Data type Earthquake holds all of the information about a single earthquake event

Constructors

Earthquake

Fields

year :: Maybe Int

month :: Maybe Int

day :: Maybe Int

place :: String

magnitude :: Maybe Double

longitude :: Maybe Double

latitude :: Maybe Double

depth :: Maybe Double

url :: String

getEarthquakeData :: String -> Earthquake

Take a single raw json string and return an earthquake, using the makeEarthquake function

makeEarthquake :: String -> String -> [Maybe Double] -> Earthquake

Takes a whole raw string, the extracted time and the extracted coordinates, and parses into an Earthquake

validateEarthquake :: Earthquake -> Bool

Take an earthquake and check it has been correctly parsed. Unparsable Json will contain a Nothing.

data Region

Data type region holds the name and boundaries of one of the world's continents

Constructors

Region

Fields

name :: String

latFrom :: Double

latTo :: Double

longFrom :: Double

longTo :: Double

makeRegion :: [String] -> Region

Make a region from a list of given strings